requirements.txt
openai>=1.0.0
pyyaml>=6.0
markdown>=3.5
requests>=2.31
gitpython>=3.1
python-dotenv>=1.0
duckduckgo-search>=4.0
jinja2>=3.1
